lua compiler: start implementing loops and if
